Make event telemetry test a sync case
